requirements.txt
openai-whisper
moviepy
torch
transformers
sentencepiece
google-cloud-texttospeech
librosa
soundfile
ffmpeg-python
opencv-python
face-alignment
numba
tqdm
requests
scipy
pathlib
git+https://github.com/openai/whisper.git

